feat!: remove password change description from /password/change
Refs: #410
