+++ src/app/(auth)/password/reset/_components/reset-password-form/use-reset-password-form.tsx
@@ -3,14 +3,14 @@ import { useCallback, useState } from 'react'
 import { useForm } from 'react-hook-form'
 import { useErrorSnackbar } from '@/app/_components/snackbars/snackbar/use-error-snackbar'
 import { useModal } from '@/components/modal/use-modal'
-import { resetPasswordSchema } from '@/schemas/request/auth'
+import { requestResetPasswordEmailSchema } from '@/schemas/request/auth'
 import { ErrorObject } from '@/types/error'
 import { resetPassword } from '@/utils/api/reset-password'
 import { HttpError } from '@/utils/error/custom/http-error'
 import type { SubmitHandler } from 'react-hook-form'
 import type { z } from 'zod'
 
-type ResetPasswordFormValues = z.infer<typeof resetPasswordSchema>
+type ResetPasswordFormValues = z.infer<typeof requestResetPasswordEmailSchema>
 
 export function useResetPasswordForm() {
   const [resultValues, setResultValues] = useState({
@@ -35,7 +35,7 @@ export function useResetPasswordForm() {
     formState: { isSubmitting, errors },
   } = useForm<ResetPasswordFormValues>({
     mode: 'onBlur',
-    resolver: zodResolver(resetPasswordSchema),
+    resolver: zodResolver(requestResetPasswordEmailSchema),
   })
 
   const handleHttpError = useCallback(
